requirements.txt
torch
torchvision
sklearn
matplotlib
pyyaml
tqdm
gdown
sklearn
argparse
pyyaml